py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/latch"]

[project]
name = "latch"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pyyaml", "fastmcp", "aiohttp", "webauthn"]

[project.scripts]
latch-hook = "latch.hook:main"
latch-serve = "latch.serve:main"
latch-dashboard = "latch.dashboard:main"
latch-enroll = "latch.enroll:main"
